Add more constants Adjust naming scheme Add types
