+++ backend/package.json
@@ -4,7 +4,7 @@
   "description": "",
   "main": "index.js",
   "scripts": {
-    "dev": "tsc && concurrently \"tsc -w\" \"nodemon ./dist/index.js\"", 
+    "dev": "tsc && concurrently \"tsc -w\" \"nodemon ./dist/index.js\"",
     "build": "tsc"
   },
   "keywords": [],
@@ -16,7 +16,10 @@
     "@types/multer": "^1.4.5",
     "@types/node": "^14.14.20",
     "@types/passport": "^1.0.5",
+    "@typescript-eslint/eslint-plugin": "^4.12.0",
+    "@typescript-eslint/parser": "^4.12.0",
     "concurrently": "^5.3.0",
+    "eslint": "^7.17.0",
     "nodemon": "^2.0.4",
     "prettier": "^2.2.1",
     "ts-node": "^9.1.1",
