+++ backend/src/config/constants.ts
@@ -4,3 +4,5 @@ export const CLIENT_SECRET =
   "SULCuWVF4E1foAy5GKm7z7hQxWUcOhLZjoBvzSdBKzN3OyNW776oaT1h69O65ozk";
 export const PORT = 3000;
 export const HOST = `http://localhost:${PORT}`;
+export const CANVAS_PORT = 8080;
+export const CANVAS = `http://localhost:${CANVAS_PORT}`;
