+++ backend/src/routes/authRoutes.ts
@@ -1,0 +1,42 @@
+import axios from "axios";
+import express, { Request, Response } from "express";
+import qs from "qs";
+import { CLIENT_ID, CLIENT_SECRET, HOST, CANVAS } from "../config/constants";
+export const authRoutes = express.Router();
+
+authRoutes.route("/init").post((req: Request, res: Response) => {
+  const redirectURL = new URL(`${CANVAS}/login/oauth2/auth`);
+
+  redirectURL.searchParams.append("client_id", CLIENT_ID);
+  redirectURL.searchParams.append("response_type", "code");
+  redirectURL.searchParams.append("redirect_uri", `${HOST}/auth/redirect`);
+  redirectURL.searchParams.append(
+    "scope",
+    "https://canvas.instructure.com/lti/feature_flags/scope/show"
+  );
+  res.redirect(redirectURL.toString());
+});
+authRoutes.route("/redirect").get((req: Request, res: Response) => {
+  const token = axios({
+    method: "post",
+    url: `${CANVAS}/login/oauth2/token`,
+    data: qs.stringify({
+      client_id: CLIENT_ID,
+      redirect_uri: `${HOST}/auth/redirect`,
+      client_secret: CLIENT_SECRET,
+      code: req.query.code,
+    }),
+    headers: {
+      "Content-Type": "application/x-www-form-urlencoded;charset=utf-8",
+    },
+  })
+    // @TODO: handle token sent in response
+    .then((res) => console.log(res))
+    .catch((err) => console.log(err));
+
+  res.status(200).redirect("complete");
+});
+
+authRoutes.route("/complete").get((req: Request, res: Response) => {
+  res.status(200).send(JSON.stringify("Hello world!"));
+});
