+++ backend/src/strategy.ts
@@ -10,6 +10,7 @@ export const strategy = new LTIStrategy(
     // https://github.com/omsmith/ims-lti#nonce-stores
     // nonceStore: new RedisNonceStore('testconsumerkey', redisClient)
   },
+  // eslint-disable-next-line @typescript-eslint/no-explicit-any
   (lti: any, done: any) => {
     // LTI launch parameters
     return done(null, "test");
